重写mem.h接口中的宏定义

+++ include/patset.h
@@ -8,8 +8,8 @@
 #define PRINT_PAT_NUM 1 /* 是否打印串长分布信息 */
 #define PRINT_PAT 0 /* 是否打印每个模式串 */
 
-typedef int64_t Pat_Num_T;
-typedef int16_t Pat_Len_T;
+typedef uint32_t Pat_Num_T;
+typedef uint8_t Pat_Len_T;
 
 #define T Patset_T
 typedef struct T *T;
@@ -20,7 +20,7 @@ struct T {
     Pat_Num_T pat_num; /* 模式串数量 */
     Pat_Len_T max_patlen; /*最大串长*/
     Pat_Len_T min_patlen; /*最小串长*/
-    int64_t total_patlen; /*串长之和*/
+    uint64_t total_patlen; /*串长之和*/
     double mean_patlen; /*平均串长*/
     double patlen_sd; /*标准差*/
     Pat_Num_T patlen_num[MAX_PAT_LEN+1]; /*串长分布，记录每一个长度含有模式串的个数*/
